DisplayResults now turns it's margin read to warn the user if target not achievable
DisplayResults.java
If the target weight is unachievable, the margin around the screen is set red
If no weights are needed to achieve the target or nearest weight to the target, then "Just the bar" is displayed

WeightsCalculator.java
Made packageName a class variable so I just need to initialize it once and reuse it
Made helper methods getWeightRef and getMultiplierRef for getting references to the weight and multiplier UI elements.  This will make saving and loading easier later and reduce duplication
In calculateWeights I no longer use an array for references to the weights and multipliers.  Use my helper methods now to get a reference

activity_display_results.xml
Put a relative layout to hold all the other UI elements inside the top level relative layout
The top level layout is the default color normally.  But can be set read to alert the user the target wasn't met as described above

strings.xmlA
Added string for "Just the bar"
Updated string for toast warning can't make target